[Feat] 임시 비밀번호 적용 API 구현 및 Test (#113)
* [Test] 임시 비밀번호 안내 이메일 발송 Test 작성

* [Feat] 임시 비밀번호 적용 API 구현 및 Test

* [Chore] testImg의 .gitkeep 삭제안되도록 After 로직 수정

+++ back/src/test/java/travelRepo/util/After.java
@@ -17,6 +17,9 @@ public class After {
         }
 
         for (File file : files) {
+            if (file.getName().equals(".gitkeep")) {
+                continue;
+            }
             file.delete();
         }
     }
